fix(output): fix combined output for new JSON output
