+++ internal/output/output.go
@@ -1,7 +1,6 @@
 package output
 
 import (
-	"encoding/json"
 	"fmt"
 	"sort"
 	"time"
@@ -74,11 +73,6 @@ type Options struct {
 	GroupKey    string
 }
 
-type ReportInput struct {
-	Metadata map[string]string
-	Root     Root
-}
-
 func outputBreakdown(resources []*schema.Resource) *Breakdown {
 	arr := make([]Resource, 0, len(resources))
 
@@ -89,7 +83,7 @@ func outputBreakdown(resources []*schema.Resource) *Breakdown {
 		arr = append(arr, outputResource(r))
 	}
 
-	sortResources(arr)
+	sortResources(arr, "")
 
 	totalMonthlyCost, totalHourlyCost := calculateTotalCosts(arr)
 
@@ -178,7 +172,7 @@ func ToOutputFormat(projects []*schema.Project) Root {
 		OnlyFields: []string{"UnsupportedCounts"},
 	})
 
-	sortResources(outResources)
+	sortResources(outResources, "")
 
 	out := Root{
 		Resources:        outResources,
@@ -190,69 +184,6 @@ func ToOutputFormat(projects []*schema.Project) Root {
 	}
 
 	return out
-}
-
-func Load(data []byte) (Root, error) {
-	var out Root
-	err := json.Unmarshal(data, &out)
-	return out, err
-}
-
-func Combine(inputs []ReportInput, opts Options) Root {
-	var combined Root
-
-	var totalHourlyCost *decimal.Decimal
-	var totalMonthlyCost *decimal.Decimal
-	summaries := make([]*ResourceSummary, 0, len(inputs))
-
-	for _, input := range inputs {
-
-		for _, r := range input.Root.Resources {
-			for k, v := range input.Metadata {
-				if r.Metadata == nil {
-					r.Metadata = make(map[string]string)
-				}
-				r.Metadata[k] = v
-			}
-
-			combined.Resources = append(combined.Resources, r)
-		}
-
-		if input.Root.TotalHourlyCost != nil {
-			if totalHourlyCost == nil {
-				totalHourlyCost = decimalPtr(decimal.Zero)
-			}
-
-			totalHourlyCost = decimalPtr(totalHourlyCost.Add(*input.Root.TotalHourlyCost))
-		}
-		if input.Root.TotalMonthlyCost != nil {
-			if totalMonthlyCost == nil {
-				totalMonthlyCost = decimalPtr(decimal.Zero)
-			}
-
-			totalMonthlyCost = decimalPtr(totalMonthlyCost.Add(*input.Root.TotalMonthlyCost))
-		}
-
-		summaries = append(summaries, input.Root.ResourceSummary)
-	}
-
-	combined.sortResources(opts.GroupKey)
-
-	combined.TotalHourlyCost = totalHourlyCost
-	combined.TotalMonthlyCost = totalMonthlyCost
-	combined.TimeGenerated = time.Now()
-	combined.ResourceSummary = combinedResourceSummaries(summaries)
-
-	return combined
-}
-
-func (r *Root) sortResources(groupKey string) {
-	sort.Slice(r.Resources, func(i, j int) bool {
-		if r.Resources[i].Metadata[groupKey] == r.Resources[j].Metadata[groupKey] {
-			return r.Resources[i].Name < r.Resources[j].Name
-		}
-		return r.Resources[i].Metadata[groupKey] < r.Resources[j].Metadata[groupKey]
-	})
 }
 
 func (r *Root) unsupportedResourcesMessage(showSkipped bool) string {
@@ -343,47 +274,6 @@ func BuildResourceSummary(resources []*schema.Resource, opts ResourceSummaryOpti
 	return s
 }
 
-func combinedResourceSummaries(summaries []*ResourceSummary) *ResourceSummary {
-	combined := &ResourceSummary{}
-
-	for _, s := range summaries {
-		if s == nil {
-			continue
-		}
-
-		combined.SupportedCounts = combineCounts(combined.SupportedCounts, s.SupportedCounts)
-		combined.UnsupportedCounts = combineCounts(combined.UnsupportedCounts, s.UnsupportedCounts)
-		combined.TotalSupported = addIntPtrs(combined.TotalSupported, s.TotalSupported)
-		combined.TotalUnsupported = addIntPtrs(combined.TotalUnsupported, s.TotalUnsupported)
-		combined.TotalNoPrice = addIntPtrs(combined.TotalNoPrice, s.TotalNoPrice)
-		combined.Total = addIntPtrs(combined.Total, s.Total)
-	}
-
-	return combined
-}
-
-func combineCounts(c1 *map[string]int, c2 *map[string]int) *map[string]int {
-	if c1 == nil && c2 == nil {
-		return nil
-	}
-
-	res := make(map[string]int)
-
-	if c1 != nil {
-		for k, v := range *c1 {
-			res[k] = v
-		}
-	}
-
-	if c2 != nil {
-		for k, v := range *c2 {
-			res[k] += v
-		}
-	}
-
-	return &res
-}
-
 func calculateTotalCosts(resources []Resource) (*decimal.Decimal, *decimal.Decimal) {
 	totalHourlyCost := decimalPtr(decimal.Zero)
 	totalMonthlyCost := decimalPtr(decimal.Zero)
@@ -409,29 +299,21 @@ func calculateTotalCosts(resources []Resource) (*decimal.Decimal, *decimal.Decim
 	return totalHourlyCost, totalMonthlyCost
 }
 
-func sortResources(resources []Resource) {
+func sortResources(resources []Resource, groupKey string) {
 	sort.Slice(resources, func(i, j int) bool {
-		return resources[i].Name < resources[j].Name
+		// If an empty group key is passed just sort by name
+		if groupKey == "" {
+			return resources[i].Name < resources[j].Name
+		}
+
+		// If the resources are in the same group then sort by name
+		if resources[i].Metadata[groupKey] == resources[j].Metadata[groupKey] {
+			return resources[i].Name < resources[j].Name
+		}
+
+		// Sort by the group key
+		return resources[i].Metadata[groupKey] < resources[j].Metadata[groupKey]
 	})
-}
-
-func addIntPtrs(i1 *int, i2 *int) *int {
-	if i1 == nil && i2 == nil {
-		return nil
-	}
-
-	val1 := 0
-	if i1 != nil {
-		val1 = *i1
-	}
-
-	val2 := 0
-	if i2 != nil {
-		val2 = *i2
-	}
-
-	res := val1 + val2
-	return &res
 }
 
 func formatAmount(d decimal.Decimal) string {
